Improve code organization and move resorce specific test classes into standalone modules.
